refactor(api): add application/certification-center-invitation directory
for controllers and routes
