+++ api/src/team/application/routes.js
@@ -1,5 +1,5 @@
-import { certificationCenterInvitationAdminRoutes } from './certification-center-invitation.admin.route.js';
-import { certificationCenterInvitationRoutes } from './certification-center-invitation.route.js';
+import { certificationCenterInvitationAdminRoutes } from './certification-center-invitation/certification-center-invitation.admin.route.js';
+import { certificationCenterInvitationRoutes } from './certification-center-invitation/certification-center-invitation.route.js';
 
 const register = async function (server) {
   server.route([...certificationCenterInvitationRoutes, ...certificationCenterInvitationAdminRoutes]);
